fix: fix http client and status monitor
