+++ hummingbot/client/hummingbot_application.py
@@ -23,7 +23,7 @@ from hummingbot.client.ui.hummingbot_cli import HummingbotCLI
 from hummingbot.client.ui.keybindings import load_key_bindings
 from hummingbot.client.ui.parser import ThrowingArgumentParser, load_parser
 from hummingbot.connector.exchange_base import ExchangeBase
-from hummingbot.connector.gateway.core import GatewayMonitor
+from hummingbot.connector.gateway.core import GatewayStatusMonitor
 from hummingbot.core.trading_core import TradingCore
 from hummingbot.core.utils.trading_pair_fetcher import TradingPairFetcher
 from hummingbot.exceptions import ArgumentParserError
@@ -76,7 +76,7 @@ class HummingbotApplication(*commands):
 
         # Script configuration support
         self.script_config: Optional[str] = None
-        self._gateway_monitor = GatewayMonitor(self)
+        self._gateway_monitor = GatewayStatusMonitor(self)
         # Start the gateway monitor in the background
         from hummingbot.core.utils.async_utils import safe_ensure_future
         safe_ensure_future(self._gateway_monitor.start())
