+++ test/hummingbot/connector/gateway/test_gateway_monitor.py
@@ -1,13 +1,13 @@
-"""Test for GatewayMonitor class."""
+"""Test for GatewayStatusMonitor class."""
 import asyncio
 import unittest
 from unittest.mock import AsyncMock, Mock, patch
 
 from hummingbot.connector.gateway.core import GatewayStatus
-from hummingbot.connector.gateway.core.gateway_monitor import GatewayMonitor
+from hummingbot.connector.gateway.core.gateway_monitor import GatewayStatusMonitor
 
 
-class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
+class TestGatewayStatusMonitor(unittest.IsolatedAsyncioTestCase):
     """Test Gateway Monitor functionality."""
 
     def setUp(self):
@@ -17,18 +17,18 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         self.mock_client_config.gateway.gateway_api_port = 15888
         self.mock_client_config.gateway.gateway_use_ssl = False
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_monitor_initialization(self, mock_gateway_client_class):
         """Test monitor initialization."""
         mock_client = AsyncMock()
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
         self.assertIsNotNone(monitor.client)
         self.assertEqual(monitor.gateway_status, GatewayStatus.OFFLINE)
         self.assertEqual(monitor.check_interval, 2.0)  # Default is 2.0, not 10.0
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_gateway_status_online(self, mock_gateway_client_class):
         """Test gateway status check when online."""
         mock_client = AsyncMock()
@@ -36,7 +36,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         mock_client.initialize_gateway = AsyncMock()
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
 
         # Test status check
         is_online = await monitor.check_once()
@@ -50,14 +50,14 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(monitor.gateway_status, GatewayStatus.ONLINE)
         mock_client.initialize_gateway.assert_called_once()
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_gateway_status_offline(self, mock_gateway_client_class):
         """Test gateway status check when offline."""
         mock_client = AsyncMock()
         mock_client.ping_gateway = AsyncMock(return_value=False)
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
 
         # Test status check
         is_online = await monitor.check_once()
@@ -66,7 +66,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         # Monitor should remain offline
         self.assertEqual(monitor.gateway_status, GatewayStatus.OFFLINE)
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_start_stop_monitoring(self, mock_gateway_client_class):
         """Test starting and stopping the monitor."""
         mock_client = AsyncMock()
@@ -74,7 +74,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         mock_client.initialize_gateway = AsyncMock()
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
         monitor.check_interval = 0.1  # Fast interval for testing
 
         # Start monitoring
@@ -89,7 +89,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         await monitor.stop()
         self.assertTrue(monitor._monitor_task.done())
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_status_transition_offline_to_online(self, mock_gateway_client_class):
         """Test transition from offline to online status."""
         mock_client = AsyncMock()
@@ -97,7 +97,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         mock_client.initialize_gateway = AsyncMock()
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
 
         # Initially offline
         is_offline = await monitor.check_once()
@@ -126,7 +126,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         # Verify initialization was called on transition
         mock_client.initialize_gateway.assert_called_once()
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_initialization_error_handling(self, mock_gateway_client_class):
         """Test error handling during gateway initialization."""
         mock_client = AsyncMock()
@@ -134,7 +134,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         mock_client.initialize_gateway = AsyncMock(side_effect=Exception("Init failed"))
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
 
         # Start monitor to trigger initialization
         await monitor.start()
@@ -145,7 +145,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(monitor.gateway_status, GatewayStatus.ONLINE)
         mock_client.initialize_gateway.assert_called_once()
 
-    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayClient')
+    @patch('hummingbot.connector.gateway.core.gateway_monitor.GatewayHttpClient')
     async def test_continuous_monitoring(self, mock_gateway_client_class):
         """Test continuous monitoring with status changes."""
         mock_client = AsyncMock()
@@ -169,7 +169,7 @@ class TestGatewayMonitor(unittest.IsolatedAsyncioTestCase):
 
         mock_gateway_client_class.get_instance.return_value = mock_client
 
-        monitor = GatewayMonitor(self.mock_client_config)
+        monitor = GatewayStatusMonitor(self.mock_client_config)
         monitor.check_interval = 0.05  # Very fast for testing
 
         # Track status changes
