+++ test_gateway_connection.py
@@ -6,7 +6,7 @@ import sys
 from pathlib import Path
 
 from hummingbot.client.config.client_config_map import ClientConfigMap
-from hummingbot.connector.gateway.core import GatewayClient
+from hummingbot.connector.gateway.core import GatewayHttpClient
 
 sys.path.insert(0, str(Path(__file__).parent))
 
@@ -16,7 +16,7 @@ async def test_gateway_connection():
     # Test with default URL
     print("Testing Gateway connection to http://localhost:15888...")
 
-    client = GatewayClient.get_instance("http://localhost:15888")
+    client = GatewayHttpClient.get_instance("http://localhost:15888")
 
     # Test ping
     print("\n1. Testing ping_gateway()...")
@@ -48,11 +48,11 @@ async def test_gateway_connection():
     config.gateway.gateway_api_host = "localhost"
     config.gateway.gateway_api_port = 15888
 
-    # Simulate what GatewayMonitor does
+    # Simulate what GatewayStatusMonitor does
     gateway_url = f"http://{config.gateway.gateway_api_host}:{config.gateway.gateway_api_port}"
     print(f"   Gateway URL: {gateway_url}")
 
-    client2 = GatewayClient.get_instance(gateway_url)
+    client2 = GatewayHttpClient.get_instance(gateway_url)
     try:
         result = await client2.ping_gateway()
         print(f"   Ping result: {result}")
